separate functions into different files, do the same for unit tests
