+++ src/commands/lambda/interfaces.ts
@@ -29,7 +29,7 @@ export interface LambdaConfigOptions {
 export interface FunctionConfiguration {
   functionARN: string
   lambdaConfig: Lambda.FunctionConfiguration
-  lambdaLibraryLayerArn: string
+  lambdaLibraryLayerArn?: string
   logGroupConfiguration?: LogGroupConfiguration
   tagConfiguration?: TagConfiguration
   updateRequest?: Lambda.UpdateFunctionConfigurationRequest
